// src/types/models.ts
export type OrderStatus =
  | 'pending_deposit'
  | 'paid'
  | 'packing'
  | 'shipping'
  | 'completed'
  | 'cancelled'